+++ relay/kusama/src/lib.rs
@@ -28,7 +28,7 @@ use frame_support::{
 	traits::{EnsureOrigin, EnsureOriginWithArg, OnRuntimeUpgrade},
 	weights::constants::{WEIGHT_PROOF_SIZE_PER_KB, WEIGHT_REF_TIME_PER_MICROS},
 };
-use kusama_runtime_constants::system_parachain::coretime::TIMESLICE_PERIOD;
+use kusama_runtime_constants::{proxy::ProxyType, system_parachain::coretime::TIMESLICE_PERIOD};
 use pallet_nis::WithMaximumOf;
 use polkadot_primitives::{
 	slashing, AccountId, AccountIndex, ApprovalVotingParams, Balance, BlockNumber, CandidateEvent,
@@ -180,7 +180,7 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
 	spec_name: create_runtime_str!("kusama"),
 	impl_name: create_runtime_str!("parity-kusama"),
 	authoring_version: 2,
-	spec_version: 1_004_000,
+	spec_version: 1_004_001,
 	impl_version: 0,
 	apis: RUNTIME_API_VERSIONS,
 	transaction_version: 26,
@@ -834,6 +834,7 @@ impl pallet_staking::Config for Runtime {
 	type EventListeners = (NominationPools, DelegatedStaking);
 	type DisablingStrategy = pallet_staking::UpToLimitDisablingStrategy;
 	type WeightInfo = weights::pallet_staking::WeightInfo<Runtime>;
+	type Filter = pallet_nomination_pools::AllPoolMembers<Runtime>;
 }
 
 impl pallet_fast_unstake::Config for Runtime {
@@ -1164,7 +1165,10 @@ parameter_types! {
 	pub const MaxPending: u16 = 32;
 }
 
-/// The type used to represent the kinds of proxying allowed.
+/// Transparent wrapper around the actual [`ProxyType`].
+///
+/// This is done to have [`ProxyType`] declared in a different crate (constants) and being able to
+/// implement [`InstanceFilter`] in this crate.
 #[derive(
 	Copy,
 	Clone,
@@ -1176,41 +1180,21 @@ parameter_types! {
 	Decode,
 	RuntimeDebug,
 	MaxEncodedLen,
-	TypeInfo,
+	Default,
 )]
-pub enum ProxyType {
-	#[codec(index = 0)]
-	Any,
-	#[codec(index = 1)]
-	NonTransfer,
-	#[codec(index = 2)]
-	Governance,
-	#[codec(index = 3)]
-	Staking,
-	// Index 4 skipped. Formerly `IdentityJudgement`.
-	#[codec(index = 5)]
-	CancelProxy,
-	#[codec(index = 6)]
-	Auction,
-	#[codec(index = 7)]
-	Society,
-	#[codec(index = 8)]
-	NominationPools,
-	#[codec(index = 9)]
-	Spokesperson,
-	#[codec(index = 10)]
-	ParaRegistration,
-}
+pub struct TransparentProxyType(ProxyType);
 
-impl Default for ProxyType {
-	fn default() -> Self {
-		Self::Any
+impl scale_info::TypeInfo for TransparentProxyType {
+	type Identity = <ProxyType as TypeInfo>::Identity;
+
+	fn type_info() -> scale_info::Type {
+		ProxyType::type_info()
 	}
 }
 
-impl InstanceFilter<RuntimeCall> for ProxyType {
+impl InstanceFilter<RuntimeCall> for TransparentProxyType {
 	fn filter(&self, c: &RuntimeCall) -> bool {
-		match self {
+		match self.0 {
 			ProxyType::Any => true,
 			ProxyType::NonTransfer => matches!(
 				c,
@@ -1316,7 +1300,7 @@ impl InstanceFilter<RuntimeCall> for ProxyType {
 		}
 	}
 	fn is_superset(&self, o: &Self) -> bool {
-		match (self, o) {
+		match (self.0, o.0) {
 			(x, y) if x == y => true,
 			(ProxyType::Any, _) => true,
 			(_, ProxyType::Any) => false,
@@ -1330,7 +1314,7 @@ impl pallet_proxy::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type RuntimeCall = RuntimeCall;
 	type Currency = Balances;
-	type ProxyType = ProxyType;
+	type ProxyType = TransparentProxyType;
 	type ProxyDepositBase = ProxyDepositBase;
 	type ProxyDepositFactor = ProxyDepositFactor;
 	type MaxProxies = MaxProxies;
@@ -1690,6 +1674,7 @@ impl pallet_nomination_pools::Config for Runtime {
 	type PalletId = PoolsPalletId;
 	type MaxPointsToBalance = MaxPointsToBalance;
 	type AdminOrigin = EitherOf<EnsureRoot<AccountId>, StakingAdmin>;
+	type Filter = pallet_staking::AllStakers<Runtime>;
 }
 
 parameter_types! {